+++ src/main/java/com/dailyon/orderservice/common/exception/CustomException.java
@@ -1,4 +1,4 @@
-package com.dailyon.orderservice.exception;
+package com.dailyon.orderservice.common.exception;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
