+++ src/main/java/com/dailyon/orderservice/domain/order/service/OrderService.java
@@ -1,7 +1,6 @@
 package com.dailyon.orderservice.domain.order.service;
 
 import com.dailyon.orderservice.domain.order.implement.OrderReader;
-import com.dailyon.orderservice.domain.order.implement.TOrderReader;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -10,7 +9,5 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class OrderService {
-    private final OrderReader orderReader;
-    private final TOrderReader tOrderReader;
-
+  private final OrderReader orderReader;
 }
