+++ src/main/java/com/dailyon/orderservice/dynamodb/implement/TOrderAppender.java
@@ -1,6 +1,5 @@
-package com.dailyon.orderservice.domain.order.implement;
+package com.dailyon.orderservice.dynamodb.implement;
 
-import com.dailyon.orderservice.domain.order.exception.OrderNotFoundException;
 import com.dailyon.orderservice.dynamodb.entity.TOrder;
 import com.dailyon.orderservice.dynamodb.repository.OrderDynamoRepository;
 import lombok.RequiredArgsConstructor;
@@ -8,10 +7,11 @@ import org.springframework.stereotype.Component;
 
 @RequiredArgsConstructor
 @Component
-public class TOrderReader {
+public class TOrderAppender {
   private final OrderDynamoRepository orderDynamoRepository;
 
-  public TOrder read(String orderId) {
-    return orderDynamoRepository.findById(orderId).orElseThrow(OrderNotFoundException::new);
+  public TOrder append(TOrder tOrder) {
+    TOrder savedOrder = orderDynamoRepository.save(tOrder);
+    return savedOrder;
   }
 }
