+++ src/main/java/com/dailyon/orderservice/dynamodb/service/request/CreateTOrderServiceRequest.java
@@ -1,0 +1,7 @@
+package com.dailyon.orderservice.dynamodb.service.request;
+
+public class CreateTOrderServiceRequest {
+    private String orderId;
+    private Long memberId;
+
+}
